requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
charset-normalizer==3.2.0
frozenlist==1.4.0
gql==3.4.1
graphql-core==3.2.3
idna==3.4
multidict==6.0.4
PyYAML==6.0.1
yarl==1.9.2
